things

+++ scripts/prottrans_seqs/prottrans_seq_gen.py
@@ -118,7 +118,7 @@ for family in pfam_seqs:
     for seq_data in sample:
         print(seq_data)
         for i in [0.25, 0.5, 0.75]:
-            number = math.ceil(len(seq_data['seq']))*i
+            number = int(math.ceil(len(seq_data['seq']))*i)
             locations = list(range(0, len(seq_data['seq'])))
             location_sample = random.sample(locations, k=number)
             new_seq = seq_data['seq']
